Gravity fix

+++ GamesProgramming/Particle.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "Particle.h"
 
-Particle::Particle(vector position, float mass, int lifespan):Shape(position, mass) {
+Particle::Particle(vector position, float mass, float lifespan):Shape(position, mass) {
 	this->lifespan = lifespan;
 	this->currLife = lifespan;
 }
