+++ GamesProgramming/Particle.h
@@ -4,7 +4,7 @@
 class Particle : public Shape
 {
 	public:
-		Particle(vector position,float mass, int life);
+		Particle(vector position,float mass, float life);
 		bool isAlive();
 		void update(int dt);
 	private:
